requirements.txt
rich==13.7.1
langchain==0.2.7
langchain-community==0.2.7
python-docx==1.1.2
python-pptx==0.6.23
openpyxl==3.1.5
PyPDF2==3.0.1
chardet==5.2.0
Markdown==3.6
beautifulsoup4==4.12.3
EbookLib==0.18
mobi==0.3.3
html2text==2024.2.26
